+++ src/kea/path.ts
@@ -5,7 +5,7 @@ export function getPathForInput(input: Input, props: Props) {
   const key = props && input.key ? input.key(props) : undefined
 
   if (input.path) {
-    return input.path(key)
+    return typeof input.path === 'function' ? input.path(key) : input.path
   }
 
   const {