+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2.1.1 - 2020-06-14
+- Read the blog post: https://kea.js.org/blog/kea-2.1
+- Simplified the syntax even more! You don't need to pass `() => ({ ... })`
+  to `actions`, `reducers`, etc if there's nothing you want from the logic.
+  Just pass `{}` instead. 
+- Simplified syntax for selectors when not using the function syntax:
+```javascript
+kea({ 
+  selectors: { 
+    doubleCounter: [
+      selectors => [selectors.counter],
+      (counter) => counter * 2
+    ]
+  }
+})
+```
+- Listeners get the store's `previousState` as their 4th argument.
+  You can use selectors (`selectors.myData(previousState)`) to get the any value as it was before the action was dispatched.
+
+## 2.0.1 - 2020-06-05
+- Fix error when calling old references to actions on unmounted logic via hooks.
+
 ## 2.0.0 - 2020-05-12
 Version 2.0 brings a lot of convenience features and one breaking change. The changes are briefly 
 described below. Read the announcement blog post to get more context: https://kea.js.org/blog/kea-2.0